Add metadata for localhost remote SP (i.e., our rails dev server)
This adds converted metadata for a SP remote defined in our Rails app. A
script has been added to the app `script/create_saml_dev_idp.rb` that
sets up the corresponding record in a developer environment.

At the moment this means you can visit `/auth/saml/fl-simplesamlphp-dev`
in our rails app, be redirected to the SimpleSAMLphp IdP, authenticate
there using `joebloggs / password` and be redirected back to the FL app,
at which point we fail to find a matching user and bomb out. Which is
exactly where we'd expect to be. Smashing.