Simplify AssertionConsumerService config
[1] says that the value of the `AssertionConsumerService` option can be
"specified in one of several endpoint formats". The endpoint formats are
listed by [2], where one of the options is a single string. A single
string can be used when "there is only a single endpoint that uses the
default binding", which I think is true in our case.

Using a single string to configure the `AssertionConsumerService` option
is a lot simpler and doesn't seem to break anything.

[1] https://simplesamlphp.org/docs/stable/simplesamlphp-reference-sp-remote
[2] https://simplesamlphp.org/docs/stable/simplesamlphp-metadata-endpoints