Populate nameid with mail attribute
We currently set the `NameIdFormat` option to the emailAddress URN. This
doesn't actually populate the `nameid` of a response with an email,
though: It only describes the format of the value.

To actually populate `nameid` with an email address, we need configure
the `simplesaml.nameidattribute` option to pull the value from the
relevant `mail` attribute.